+++ NEWS.md
@@ -7,6 +7,7 @@
 * make_forest_plot() no longer returns data. But the plot data is available at .$plot$data
 * New argument envir added to make_shape_plot (allows the user to specify the environment for evaluating the plot code).
 * ggtext::element_markdown() now used for y-axis labels in make_forest_plot()
+* fix_panel_width() replaced by fix_panel()
 * Improved handling of unicode characters.
 * Updated documentation.
 * Fixes and internal code improvements.